// app/src/db/models/transcripts.ts
import mongoose from "mongoose";


/*
{message_id: "496c0f70-e5a6-4d3e-99c2-7d0e1cde3d7d", text: "Hello, how are you?", type: "processed",…}
message_id
:
"496c0f70-e5a6-4d3e-99c2-7d0e1cde3d7d"
speaker
:
"SPEAKER_00"
text
:
"Hello, how are you?"
type
:
"processed"
 */

const transcriptsSchema = new mongoose.Schema({
    session: { type: mongoose.Schema.Types.ObjectId, ref: "sessions", required: true, index: true},
    message_id: { type: String, required: true, index: true }, // id of message, generated by stt service
    speaker: { type: String, required: true },

    raw_text: { type: String, required: true },

    processed_text: String,

    corrected_text: String,
    concise_text: String,
    highlighted_text: String,
}, {
    timestamps: true
});

const Transcripts = mongoose.model("transcripts", transcriptsSchema);

export default Transcripts;